perf: move client and server on one binary
