+++ perf/src/lib.rs
@@ -14,12 +14,14 @@ use quinn::{
 use rustls::crypto::ring::cipher_suite;
 use socket2::{Domain, Protocol, Socket, Type};
 use tracing::warn;
-use tracing_subscriber::{EnvFilter, fmt, layer::SubscriberExt, util::SubscriberInitExt};
 
 #[cfg_attr(not(feature = "json-output"), allow(dead_code))]
 pub mod stats;
 
 pub mod noprotection;
+
+pub mod client;
+pub mod server;
 
 // Common options between client and server binary
 #[derive(Parser)]
@@ -210,17 +212,6 @@ impl CongestionAlgorithm {
             CongestionAlgorithm::NewReno => Arc::new(congestion::NewRenoConfig::default()),
         }
     }
-}
-
-pub fn init_tracing() {
-    tracing_subscriber::registry()
-        .with(
-            EnvFilter::try_from_default_env()
-                .or_else(|_| EnvFilter::try_new("warn"))
-                .unwrap(),
-        )
-        .with(fmt::layer())
-        .init();
 }
 
 pub static PERF_CIPHER_SUITES: &[rustls::SupportedCipherSuite] = &[
